conscience-it-your-inner-guide-main/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic==2.5.0
cryptography==41.0.7
websockets==12.0
openai==1.52.0
